services/extraction/requirements.txt
pika==1.3.2
PyPDF2==3.0.1
python-magic==0.4.27
pdfplumber==0.10.3
python-dotenv==1.0.0
Flask==3.0.0